🚧 Rework Windows distribution workflow
*   `.github/workflows/build_and_release.yml`:
    *   Disabled the `publish-windows` job.
    *   Updated `flutter_distributor` commands to `fastforge`.
    *   Set GitHub releases to `draft`.
*   `gui/distribute_options.yaml`:
    *   Switched Windows packaging target from `exe` to `msix`.
*   `gui/windows/packaging/exe/make_config.yaml`:
    *   Removed `install_dir_name` field.

This refactors the Windows distribution process. The automated Microsoft Store publishing job was disabled due to a locked Microsoft tenant. The workflow will now build the MSIX package, which can then be published manually. `flutter_distributor` was renamed to `fastforge`. Releases are now drafted to allow for manual review and management of the MSIX artifact. The `exe` packaging target is deprecated in favor of `msix`.